Move attribute utils from logs to common, remove unused headers
